# src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/product-types": {
      "get": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "getAllProductTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "updateProductType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "createProductType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "getAllCustomers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "createCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/api/bankproducts": {
      "get": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "getAllBankProducts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntityResponseBankProduct"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "createBankProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BankProduct"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBankProduct"
                }
              }
            }
          }
        }
      }
    },
    "/api/bankproducts/{productId}/transaction": {
      "post": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "makeTransaction",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBankProduct"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/product-types/{id}": {
      "get": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "getProductTypeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product-type-controller"
        ],
        "operationId": "deleteProductType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/bankproducts/{productId}": {
      "get": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "getBankProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBankProduct"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "deleteBankProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/bankproducts/{productId}/transactions": {
      "get": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "getProductTransactions",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseListTransaction"
                }
              }
            }
          }
        }
      }
    },
    "/api/bankproducts/{productId}/balance": {
      "get": {
        "tags": [
          "bank-product-controller"
        ],
        "operationId": "getProductBalance",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBigDecimal"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "commission": {
            "type": "number",
            "format": "double"
          },
          "transactionCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "documentNumber": {
            "type": "string"
          },
          "customerType": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "AccountHolder": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "isHolder": {
            "type": "boolean"
          }
        }
      },
      "BankProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productNumber": {
            "type": "string"
          },
          "typeProductId": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "accountHolders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountHolder"
            }
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "substract": {
            "type": "boolean"
          },
          "amount": {
            "type": "number"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "ResponseBankProduct": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/BankProduct"
          }
        }
      },
...
      "ContentDisposition": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "charset": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "readDate": {
            "type": "string",
            "format": "date-time",
            "deprecated": true
          },
          "inline": {
            "type": "boolean"
          },
          "formData": {
            "type": "boolean"
          },
          "attachment": {
            "type": "boolean"
          }
        }
      },
      "HttpMethod": {
        "type": "object"
      },
      "HttpRange": {
        "type": "object"
      },
      "HttpStatusCode": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "is1xxInformational": {
            "type": "boolean"
          },
          "is2xxSuccessful": {
            "type": "boolean"
          },
          "is3xxRedirection": {
            "type": "boolean"
          },
          "is5xxServerError": {
            "type": "boolean"
          },
          "is4xxClientError": {
            "type": "boolean"
          }
        }
      },
      "MediaType": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "subtype": {
            "type": "string"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "qualityValue": {
            "type": "number",
            "format": "double"
          },
          "concrete": {
            "type": "boolean"
          },
          "wildcardSubtype": {
            "type": "boolean"
          },
          "subtypeSuffix": {
            "type": "string"
          },
          "wildcardType": {
            "type": "boolean"
          },
          "charset": {
            "type": "string"
          }
        }
      },
      "ResponseEntityResponseBankProduct": {
        "type": "object",
        "properties": {
          "headers": {
            "type": "object",
            "properties": {
              "empty": {
                "type": "boolean"
              },
              "location": {
                "type": "string",
                "format": "uri"
              },
              "host": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "object",
                    "properties": {
                      "hostAddress": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string",
                        "format": "byte"
                      },
                      "hostName": {
                        "type": "string"
                      },
                      "linkLocalAddress": {
                        "type": "boolean"
                      },
                      "multicastAddress": {
                        "type": "boolean"
                      },
                      "anyLocalAddress": {
                        "type": "boolean"
                      },
                      "loopbackAddress": {
                        "type": "boolean"
                      },
                      "siteLocalAddress": {
                        "type": "boolean"
                      },
                      "mcglobal": {
                        "type": "boolean"
                      },
                      "mcnodeLocal": {
                        "type": "boolean"
                      },
                      "mclinkLocal": {
                        "type": "boolean"
                      },
                      "mcsiteLocal": {
                        "type": "boolean"
                      },
                      "mcorgLocal": {
                        "type": "boolean"
                      },
                      "canonicalHostName": {
                        "type": "string"
                      }
                    }
                  },
                  "port": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "unresolved": {
                    "type": "boolean"
                  },
                  "hostName": {
                    "type": "string"
                  },
                  "hostString": {
                    "type": "string"
                  }
                }
              },
              "all": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "writeOnly": true
              },
              "lastModified": {
                "type": "integer",
                "format": "int64"
              },
              "date": {
                "type": "integer",
                "format": "int64"
              },
              "contentLength": {
                "type": "integer",
                "format": "int64"
              },
              "range": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpRange"
                }
              },
              "accept": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MediaType"
                }
              },
              "allow": {
                "uniqueItems": true,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpMethod"
                }
              },
              "acceptCharset": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "acceptLanguage": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "type": "string"
                    },
                    "weight": {
                      "type": "number",
                      "format": "double"
                    }
                  }
                }
              },
              "basicAuth": {
                "type": "string",
                "writeOnly": true
              },
              "acceptPatch": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MediaType"
                }
              },
              "connection": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "contentLanguage": {
                "type": "object",
                "properties": {
                  "language": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "variant": {
                    "type": "string"
                  },
                  "script": {
                    "type": "string"
                  },
                  "unicodeLocaleAttributes": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "unicodeLocaleKeys": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "displayLanguage": {
                    "type": "string"
                  },
                  "displayScript": {
                    "type": "string"
                  },
                  "displayCountry": {
                    "type": "string"
                  },
                  "displayVariant": {
                    "type": "string"
                  },
                  "extensionKeys": {
                    "uniqueItems": true,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "iso3Language": {
                    "type": "string"
                  },
                  "iso3Country": {
                    "type": "string"
                  }
                }
              },
              "etag": {
                "type": "string"
              },
              "cacheControl": {
                "type": "string"
              },
              "bearerAuth": {
                "type": "string",
                "writeOnly": true
              },
              "vary": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "upgrade": {
                "type": "string"
              },
              "origin": {
                "type": "string"
              },
              "ifMatch": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "ifNoneMatch": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "pragma": {
                "type": "string"
              },
              "expires": {
                "type": "integer",
                "format": "int64"
              },
              "accessControlAllowCredentials": {
                "type": "boolean"
              },
              "acceptLanguageAsLocales": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "language": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "variant": {
                      "type": "string"
                    },
                    "script": {
                      "type": "string"
                    },
                    "unicodeLocaleAttributes": {
                      "uniqueItems": true,
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "unicodeLocaleKeys": {
                      "uniqueItems": true,
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "displayLanguage": {
                      "type": "string"
                    },
                    "displayScript": {
                      "type": "string"
                    },
                    "displayCountry": {
                      "type": "string"
                    },
                    "displayVariant": {
                      "type": "string"
                    },
                    "extensionKeys": {
                      "uniqueItems": true,
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "iso3Language": {
                      "type": "string"
                    },
                    "iso3Country": {
                      "type": "string"
                    }
                  }
                }
              },
              "accessControlAllowHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "contentDisposition": {
                "$ref": "#/components/schemas/ContentDisposition"
              },
              "accessControlRequestMethod": {
                "$ref": "#/components/schemas/HttpMethod"
              },
              "accessControlAllowMethods": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HttpMethod"
                }
              },
              "accessControlRequestHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "accessControlMaxAge": {
                "type": "integer",
                "format": "int64"
              },
              "ifUnmodifiedSince": {
                "type": "integer",
                "format": "int64"
              },
              "accessControlAllowOrigin": {
                "type": "string"
              },
              "accessControlExposeHeaders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "contentType": {
                "$ref": "#/components/schemas/MediaType"
              },
              "ifModifiedSince": {
                "type": "integer",
                "format": "int64"
              }
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "body": {
            "$ref": "#/components/schemas/ResponseBankProduct"
          },
          "statusCodeValue": {
            "type": "integer",
            "format": "int32",
            "deprecated": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        }
      },
      "ResponseListTransaction": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        }
      },
      "ResponseBigDecimal": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "number"
          }
        }
      },
      "ResponseObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  }
}